+++ mappi/include/Objects/State.hpp
@@ -18,10 +18,10 @@ namespace mappi::objects{
 
 struct Odometry2d
 {
-    float x, y, yaw;
-    float vx, vy, wz;
-    float steering;
-    float stamp;
+    double x, y, yaw;
+    double vx, vy, wz;
+    double steering;
+    double stamp;
 
     /**
      * @brief Construct a new Odometry 2d object
@@ -34,13 +34,13 @@ struct Odometry2d
 
 struct State
 {
-  xt::xtensor<float, 2> vx;
-  xt::xtensor<float, 2> vy;
-  xt::xtensor<float, 2> wz;
+  xt::xtensor<double, 2> vx;
+  xt::xtensor<double, 2> vy;
+  xt::xtensor<double, 2> wz;
 
-  xt::xtensor<float, 2> cvx;
-  xt::xtensor<float, 2> cvy;
-  xt::xtensor<float, 2> cwz;
+  xt::xtensor<double, 2> cvx;
+  xt::xtensor<double, 2> cvy;
+  xt::xtensor<double, 2> cwz;
 
   Odometry2d odom;
 
@@ -52,13 +52,13 @@ struct State
    */
   void reset(unsigned int batch_size, unsigned int time_steps)
   {
-    vx = xt::zeros<float>({batch_size, time_steps});
-    vy = xt::zeros<float>({batch_size, time_steps});
-    wz = xt::zeros<float>({batch_size, time_steps});
+    vx = xt::zeros<double>({batch_size, time_steps});
+    vy = xt::zeros<double>({batch_size, time_steps});
+    wz = xt::zeros<double>({batch_size, time_steps});
 
-    cvx = xt::zeros<float>({batch_size, time_steps});
-    cvy = xt::zeros<float>({batch_size, time_steps});
-    cwz = xt::zeros<float>({batch_size, time_steps});
+    cvx = xt::zeros<double>({batch_size, time_steps});
+    cvy = xt::zeros<double>({batch_size, time_steps});
+    cwz = xt::zeros<double>({batch_size, time_steps});
   }
 };
