+++ mappi/include/Objects/Trajectory.hpp
@@ -23,9 +23,9 @@ namespace mappi::objects {
 
 struct Trajectory
 {
-    xt::xtensor<float, 2> x;
-    xt::xtensor<float, 2> y;
-    xt::xtensor<float, 2> yaw;
+    xt::xtensor<double, 2> x;
+    xt::xtensor<double, 2> y;
+    xt::xtensor<double, 2> yaw;
 
     /**
      * @brief Resets trajectory
@@ -35,18 +35,18 @@ struct Trajectory
      */
     void reset(unsigned int batch_size, unsigned int time_steps)
     {
-        x = xt::zeros<float>({batch_size, time_steps});
-        y = xt::zeros<float>({batch_size, time_steps});
-        yaw = xt::zeros<float>({batch_size, time_steps});
+        x = xt::zeros<double>({batch_size, time_steps});
+        y = xt::zeros<double>({batch_size, time_steps});
+        yaw = xt::zeros<double>({batch_size, time_steps});
     }
 
 };
 
 struct Path
 {
-    xt::xtensor<float, 1> x;
-    xt::xtensor<float, 1> y;
-    xt::xtensor<float, 1> yaw;
+    xt::xtensor<double, 1> x;
+    xt::xtensor<double, 1> y;
+    xt::xtensor<double, 1> yaw;
     xt::xtensor<bool, 1> free;
 
     /**
@@ -56,9 +56,9 @@ struct Path
      */
     void reset(unsigned int size)
     {
-        x = xt::zeros<float>({size});
-        y = xt::zeros<float>({size});
-        yaw = xt::zeros<float>({size});
+        x = xt::zeros<double>({size});
+        y = xt::zeros<double>({size});
+        yaw = xt::zeros<double>({size});
         free = xt::ones<bool>({size});
     }
